+++ src/PIL/DdsImagePlugin.py
@@ -480,7 +480,8 @@ class DdsImageFile(ImageFile.ImageFile):
 class DdsRgbDecoder(ImageFile.PyDecoder):
     _pulls_fd = True
 
-    def decode(self, buffer):
+    def decode(self, buffer: bytes) -> tuple[int, int]:
+        assert self.fd is not None
         bitcount, masks = self.args
 
         # Some masks will be padded with zeros, e.g. R 0b11 G 0b1100
@@ -511,7 +512,7 @@ class DdsRgbDecoder(ImageFile.PyDecoder):
         return -1, 0
 
 
-def _save(im: Image.Image, fp: IO[bytes], filename: str) -> None:
+def _save(im: Image.Image, fp: IO[bytes], filename: str | bytes) -> None:
     if im.mode not in ("RGB", "RGBA", "L", "LA"):
         msg = f"cannot write mode {im.mode} as DDS"
         raise OSError(msg)